requirements.txt
fastapi
uvicorn[standard]
telethon
sqlalchemy
psycopg2-binary
pydantic
python-dotenv